- atualiza os controllers e services de Avaliacao e Comentario

+++ src/main/java/br/com/ifba/bookboxd/avaliacao/service/AvaliacaoService.java
@@ -1,11 +1,13 @@
 package br.com.ifba.bookboxd.avaliacao.service;
 
-import br.com.ifba.bookboxd.avaliacao.service.AvaliacaoIService;
 import br.com.ifba.bookboxd.avaliacao.entity.Avaliacao;
 import br.com.ifba.bookboxd.comentario.entity.Comentario;
 import br.com.ifba.bookboxd.livro.entity.Livro;
 import br.com.ifba.bookboxd.usuario.entity.Usuario;
 import br.com.ifba.bookboxd.avaliacao.repository.AvaliacaoRepository;
+import br.com.ifba.bookboxd.infrastruture.util.StringUtil;
+import br.com.ifba.bookboxd.livro.repository.LivroRepository;
+import br.com.ifba.bookboxd.usuario.repository.UsuarioRepository;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -18,22 +20,80 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class AvaliacaoService implements AvaliacaoIService {
     
-    private final AvaliacaoRepository avaliacaoRepository; 
+    private final AvaliacaoRepository avaliacaoRepository;
+    private final UsuarioRepository usuarioRepository;
+    private final LivroRepository livroRepository; 
+    
+    private void validarAvaliacao(Avaliacao avaliacao){
+        if(avaliacao == null){
+            throw new RuntimeException("Dados de avaliação não preenchidos");
+        }
+        if(avaliacao.getUsuario() == null){
+            throw new RuntimeException("A avaliação precisa estar vinculada a um Usuario");
+        }
+        if(avaliacao.getLivro() == null){
+            throw new RuntimeException("A avaliação precisa estar vinculada a um Livro");
+        }
+        if(avaliacao.getNota() < 0 || avaliacao.getNota() > 5){
+            throw new RuntimeException("Nota precisar está entre 0 e 5");
+        }
+        if(StringUtil.isEmpty(avaliacao.getAvaliacao())){
+            throw new RuntimeException("Avaliação não preenchida");
+        }
+    }
+    
+    private void validarId(Long id){
+        if(!StringUtil.isIdValido(id)){
+            throw new RuntimeException("Id inválido");
+        }
+    }
     
     @Override
     public Avaliacao save(Avaliacao avaliacao) {
+        validarAvaliacao(avaliacao);
+        if(avaliacao.getId() != null){
+            throw new RuntimeException("Avaliação ja existe no banco de dados");
+        }
+        if(avaliacao.getDataPublicacao() == null){
+            avaliacao.setDataPublicacao(LocalDate.now());
+        }
+        
         log.info("Salvando avaliação do livro ID: {}", avaliacao.getLivro().getId());
         return avaliacaoRepository.save(avaliacao);
     }
-
+    
+    @Override
+    public Avaliacao update(Avaliacao avaliacao) {
+        validarAvaliacao(avaliacao);
+        validarId(avaliacao.getId());
+        
+        if(!avaliacaoRepository.existsById(avaliacao.getId())){
+            throw new RuntimeException("Avaliação não encontrado com id: " + avaliacao.getId());
+        }
+        
+        log.info("Atualizando avaliação do livro ID: {}", avaliacao.getLivro().getId());
+        return avaliacaoRepository.save(avaliacao);
+    }
+    
     @Override
     public Optional<Avaliacao> findById(Long id) {
+        validarId(id);
         log.info("Buscando avaliação por ID: {}", id);
-        return avaliacaoRepository.findById(id);
+        Optional<Avaliacao> avaliacao = avaliacaoRepository.findById(id);
+        
+        if(avaliacao.isEmpty()){
+            throw new RuntimeException("Avaliacao não encontrado com id: " + id);
+        }
+        return avaliacao;
     }
 
     @Override
     public void delete(Long id) {
+        validarId(id);
+        
+        if(!avaliacaoRepository.existsById(id)){
+            throw new RuntimeException("Avaliacao não encontradoa com id: " + id);
+        }
         log.info("Deletando avaliação com ID: {}", id);
         avaliacaoRepository.deleteById(id);
     }
@@ -41,60 +101,100 @@ public class AvaliacaoService implements AvaliacaoIService {
     @Override
     public List<Avaliacao> findAll() {
         log.info("Listando todas as avaliações");
-        return avaliacaoRepository.findAll();
+        
+        List<Avaliacao> avaliacoes = avaliacaoRepository.findAll();
+        
+        if(avaliacoes.isEmpty()){
+            throw new RuntimeException("Nenhuma avaliacao encontrada");
+        }
+        return avaliacoes;
     }
 
     @Override
     public List<Avaliacao> findByLivroId(Long livroId) {
+        validarId(livroId);
         log.info("Buscando avaliações do livro ID: {}", livroId);
-        return avaliacaoRepository.findByLivroId(livroId);
+        List<Avaliacao> avaliacoes = avaliacaoRepository.findByLivroId(livroId);
+        
+        if(avaliacoes.isEmpty()){
+            throw new RuntimeException("Nenhuma avaliação encontrada para o livro id: " + livroId);
+        }
+        return avaliacoes;
     }
 
     @Override
     public List<Avaliacao> findByUsuarioId(Long usuarioId) {
+        validarId(usuarioId);
         log.info("Buscando avaliações do usuário ID: {}", usuarioId);
-        return avaliacaoRepository.findByUsuarioId(usuarioId);
+        List<Avaliacao> avaliacoes = avaliacaoRepository.findByUsuarioId(usuarioId);
+        
+        if(avaliacoes.isEmpty()){
+            throw new RuntimeException("Nenhuma avaliação encontrada para o usuário id: " + usuarioId);
+        }
+        return avaliacoes;
     }
 
     @Override
     public void editarTexto(Long avaliacaoId, String novoTexto) {
+        validarId(avaliacaoId);
+        if(StringUtil.isEmpty(novoTexto)){
+            throw new RuntimeException("O novo texto da avaliação não pode ser vazio");
+        }
+        
+        Avaliacao avaliacao = avaliacaoRepository.findById(avaliacaoId)
+                .orElseThrow(() -> new RuntimeException("Avaliação não encontrada com id: " + avaliacaoId));
+        
         log.info("Editando texto da avaliação ID: {}", avaliacaoId);
-        avaliacaoRepository.findById(avaliacaoId).ifPresent(avaliacao -> {
-            avaliacao.editarTexto(novoTexto);
-            avaliacaoRepository.save(avaliacao);
-        });
+        avaliacao.editarTexto(novoTexto);
+        avaliacaoRepository.save(avaliacao);
     }
 
     @Override
-    public Comentario adicionarComentario(Long avaliacaoId, Usuario autor, String texto) {
+    public Comentario adicionarComentario(Long avaliacaoId, Long autorId, String texto) {
+        validarId(avaliacaoId);
+        validarId(autorId);
+        
+        if(StringUtil.isEmpty(texto)){
+            throw new RuntimeException("O texto do comentário não pode ser vazio");
+        }
+        
+        Avaliacao avaliacao = avaliacaoRepository.findById(avaliacaoId)
+                .orElseThrow(() -> new RuntimeException("Avaliação não encontrada com id: " + avaliacaoId));
+        
+        Usuario autor = usuarioRepository.findById(autorId)
+                .orElseThrow(() -> new RuntimeException("Usuário (autor do comentário) não encontrado com id: " + autorId));
+        
         log.info("Adicionando comentario na avaliação ID: {}", avaliacaoId);
-        return avaliacaoRepository.findById(avaliacaoId).map(avaliacao -> {
-            Comentario comentario = avaliacao.adicionarComentario(autor, texto);
-            avaliacaoRepository.save(avaliacao);
-            return comentario;
-        }).orElseThrow(() -> new RuntimeException("Avaliação não encontrada: " + avaliacaoId));
+        Comentario comentario = avaliacao.adicionarComentario(autor, texto);
+        avaliacaoRepository.save(avaliacao);
+        return comentario;
     }
 
     @Override
     public void toggleSpoiler(Long avaliacaoId) {
+        validarId(avaliacaoId);
+        
+        Avaliacao avaliacao = avaliacaoRepository.findById(avaliacaoId)
+                .orElseThrow(() -> new RuntimeException("Avaliação não encontrada com id: " + avaliacaoId));
+        
         log.info("Alternando spoiler da avaliação ID: {}", avaliacaoId);
-        avaliacaoRepository.findById(avaliacaoId).ifPresent(avaliacao -> {
-            avaliacao.toggleSpoiler();
-            avaliacaoRepository.save(avaliacao);
-        });
+        avaliacao.toggleSpoiler();
+        avaliacaoRepository.save(avaliacao);
     } 
     
     @Override
-    public Avaliacao criarAvaliacao(Usuario usuario, Livro livro, int nota, String comentario, boolean contemSpoiler) {
+    public Avaliacao criarAvaliacao(Long usuarioId, Long livroId, int nota, String texto, boolean contemSpoiler) {
+        validarId(usuarioId);
+        validarId(livroId);
         
-        Avaliacao avaliacao = new Avaliacao();
-        avaliacao.setUsuario(usuario);
-        avaliacao.setLivro(livro);
-        avaliacao.setNota(nota);
-        avaliacao.setComentario(comentario);
-        avaliacao.setContemSpoiler(contemSpoiler);
-        avaliacao.setDataPublicacao(LocalDate.EPOCH.now());
+        Usuario usuario = usuarioRepository.findById(usuarioId)
+                .orElseThrow(() -> new RuntimeException("Usuário não encontrado com id: " + usuarioId));
 
-        return avaliacaoRepository.save(avaliacao);
+        Livro livro = livroRepository.findById(livroId)
+                .orElseThrow(() -> new RuntimeException("Livro não encontrado com id: " + livroId));
+        
+        Avaliacao avaliacao = usuario.criarAvaliacao(livro, nota, texto, contemSpoiler);
+        
+        return save(avaliacao);
     }
 }
